Address review comments
This commit splits out some methods from Files into FilesSupport
in order to reduce the raw line count in Files.  It also changes
the handling of Thread interruptions received during the retry
waits for rename and delete operations -- remember and re-assert.
There are also some minor changes to Shell.
